Create a clustering subpackage
