+++ tslearn/clustering/kmeans.py
@@ -1,11 +1,3 @@
-"""
-The :mod:`tslearn.clustering` module gathers time series specific clustering
-algorithms.
-
-**User guide:** See the :ref:`Clustering <clustering>` section for further 
- details.
-"""
-
 from sklearn.base import ClusterMixin, TransformerMixin
 from sklearn.metrics.pairwise import pairwise_kernels
 try:
@@ -14,8 +6,6 @@ try:
 except ImportError:
     # Deprecated from sklearn v0.24 onwards
     from sklearn.cluster.k_means_ import _k_init
-from sklearn.metrics.cluster import \
-    silhouette_score as sklearn_silhouette_score
 from sklearn.utils import check_random_state
 from sklearn.utils.extmath import stable_cumsum
 from sklearn.utils.validation import _check_sample_weight
@@ -23,18 +13,18 @@ from scipy.spatial.distance import cdist
 import numpy
 import warnings
 
-from tslearn.metrics import cdist_gak, cdist_dtw, cdist_soft_dtw, \
-    cdist_soft_dtw_normalized, sigma_gak
+from tslearn.metrics import cdist_gak, cdist_dtw, cdist_soft_dtw, sigma_gak
 from tslearn.barycenters import euclidean_barycenter, \
     dtw_barycenter_averaging, softdtw_barycenter
 from sklearn.utils import check_array
 from sklearn.utils.validation import check_is_fitted
-from tslearn.preprocessing import TimeSeriesScalerMeanVariance, \
-    TimeSeriesResampler
-from tslearn.utils import (to_time_series_dataset, to_time_series,
-                           to_sklearn_dataset, check_dims)
-from tslearn.cycc import cdist_normalized_cc, y_shifted_sbd_vec
+from tslearn.utils import (to_time_series_dataset, to_sklearn_dataset,
+                           check_dims)
 from tslearn.bases import BaseModelPackage, TimeSeriesBaseEstimator
+
+from .utils import (EmptyClusterError, _check_initial_guess,
+                    _check_no_empty_cluster, _check_full_length,
+                    TimeSeriesCentroidBasedClusteringMixin, _compute_inertia)
 
 __author__ = 'Romain Tavenard romain.tavenard[at]univ-rennes2.fr'
 # Kernel k-means is derived from https://gist.github.com/mblondel/6230787 by
@@ -125,200 +115,6 @@ def _k_init_metric(X, n_clusters, cdist_metric, random_state,
         centers[c] = X[best_candidate]
 
     return centers
-
-
-class EmptyClusterError(Exception):
-    def __init__(self, message=""):
-        super().__init__()
-        self.message = message
-
-    def __str__(self):
-        if len(self.message) > 0:
-            suffix = " (%s)" % self.message
-        else:
-            suffix = ""
-        return "Cluster assignments lead to at least one empty cluster" + \
-               suffix
-
-
-def _check_no_empty_cluster(labels, n_clusters):
-    """Check that all clusters have at least one sample assigned.
-    """
-
-    for k in range(n_clusters):
-        if numpy.sum(labels == k) == 0:
-            raise EmptyClusterError
-
-
-def _check_full_length(centroids):
-    """Check that provided centroids are full-length (ie. not padded with
-    nans).
-
-    If some centroids are found to be padded with nans, TimeSeriesResampler is
-    used to resample the centroids.
-    """
-    resampler = TimeSeriesResampler(sz=centroids.shape[1])
-    return resampler.fit_transform(centroids)
-
-
-def _compute_inertia(distances, assignments, squared=True):
-    """Derive inertia (average of squared distances) from pre-computed
-    distances and assignments.
-
-    Examples
-    --------
-    >>> dists = numpy.array([[1., 2., 0.5], [0., 3., 1.]])
-    >>> assign = numpy.array([2, 0])
-    >>> _compute_inertia(dists, assign)
-    0.125
-    """
-    n_ts = distances.shape[0]
-    if squared:
-        return numpy.sum(distances[numpy.arange(n_ts),
-                                   assignments] ** 2) / n_ts
-    else:
-        return numpy.sum(distances[numpy.arange(n_ts), assignments]) / n_ts
-
-
-def silhouette_score(X, labels, metric=None, sample_size=None,
-                     metric_params=None, n_jobs=None, verbose=0, 
-                     random_state=None, **kwds):
-    """Compute the mean Silhouette Coefficient of all samples (cf.  [1]_ and
-    [2]_).
-
-    Read more in the `scikit-learn documentation
-    <http://scikit-learn.org/stable/modules/clustering.html\
-    #silhouette-coefficient>`_.
-
-    Parameters
-    ----------
-    X : array [n_ts, n_ts] if metric == "precomputed", or, \
-             [n_ts, sz, d] otherwise
-        Array of pairwise distances between time series, or a time series
-        dataset.
-    labels : array, shape = [n_ts]
-         Predicted labels for each time series.
-    metric : string, callable or None (default: None)
-        The metric to use when calculating distance between time series.
-        Should be one of {'dtw', 'softdtw', 'euclidean'} or a callable distance
-        function or None.
-        If 'softdtw' is passed, a normalized version of Soft-DTW is used that
-        is defined as `sdtw_(x,y) := sdtw(x,y) - 1/2(sdtw(x,x)+sdtw(y,y))`.
-        If X is the distance array itself, use ``metric="precomputed"``.
-        If None, dtw is used.
-    sample_size : int or None (default: None)
-        The size of the sample to use when computing the Silhouette Coefficient
-        on a random subset of the data.
-        If ``sample_size is None``, no sampling is used.
-    metric_params : dict or None (default: None)
-        Parameter values for the chosen metric.
-        For metrics that accept parallelization of the cross-distance matrix
-        computations, `n_jobs` key passed in `metric_params` is overridden by
-        the `n_jobs` argument.
-
-    n_jobs : int or None, optional (default=None)
-        The number of jobs to run in parallel for cross-distance matrix
-        computations.
-        Ignored if the cross-distance matrix cannot be computed using
-        parallelization.
-        ``None`` means 1 unless in a :obj:`joblib.parallel_backend` context.
-        ``-1`` means using all processors. See scikit-learns'
-        `Glossary <https://scikit-learn.org/stable/glossary.html#term-n-jobs>`_
-        for more details.
-
-    verbose : int (default: 0)
-        If nonzero, print information about the inertia while learning
-        the model and joblib progress messages are printed.  
-
-    random_state : int, RandomState instance or None, optional (default: None)
-        The generator used to randomly select a subset of samples.  If int,
-        random_state is the seed used by the random number generator; If
-        RandomState instance, random_state is the random number generator; If
-        None, the random number generator is the RandomState instance used by
-        `np.random`. Used when ``sample_size is not None``.
-    **kwds : optional keyword parameters
-        Any further parameters are passed directly to the distance function,
-        just as for the `metric_params` parameter.
-
-    Returns
-    -------
-    silhouette : float
-        Mean Silhouette Coefficient for all samples.
-
-    References
-    ----------
-    .. [1] `Peter J. Rousseeuw (1987). "Silhouettes: a Graphical Aid to the
-       Interpretation and Validation of Cluster Analysis". Computational
-       and Applied Mathematics 20: 53-65.
-       <http://www.sciencedirect.com/science/article/pii/0377042787901257>`_
-    .. [2] `Wikipedia entry on the Silhouette Coefficient
-           <https://en.wikipedia.org/wiki/Silhouette_(clustering)>`_
-
-    Examples
-    --------
-    >>> from tslearn.generators import random_walks
-    >>> from tslearn.metrics import cdist_dtw
-    >>> numpy.random.seed(0)
-    >>> X = random_walks(n_ts=20, sz=16, d=1)
-    >>> labels = numpy.random.randint(2, size=20)
-    >>> silhouette_score(X, labels, metric="dtw")  # doctest: +ELLIPSIS
-    0.13383800...
-    >>> silhouette_score(X, labels, metric="euclidean")  # doctest: +ELLIPSIS
-    0.09126917...
-    >>> silhouette_score(X, labels, metric="softdtw")  # doctest: +ELLIPSIS
-    0.17953934...
-    >>> silhouette_score(X, labels, metric="softdtw",
-    ...                  metric_params={"gamma": 2.}) \
-    # doctest: +ELLIPSIS
-    0.17591060...
-    >>> silhouette_score(cdist_dtw(X), labels,
-    ...                  metric="precomputed")  # doctest: +ELLIPSIS
-    0.13383800...
-    """
-    sklearn_metric = None
-    if metric_params is None:
-        metric_params_ = {}
-    else:
-        metric_params_ = metric_params.copy()
-    for k in kwds.keys():
-        metric_params_[k] = kwds[k]
-    if "n_jobs" in metric_params_.keys():
-        del metric_params_["n_jobs"]
-    if metric == "precomputed":
-        sklearn_X = X
-    elif metric == "dtw" or metric is None:
-        sklearn_X = cdist_dtw(X, n_jobs=n_jobs, verbose=verbose,
-                              **metric_params_)
-    elif metric == "softdtw":
-        sklearn_X = cdist_soft_dtw_normalized(X, **metric_params_)
-    elif metric == "euclidean":
-        X_ = to_time_series_dataset(X)
-        X_ = X_.reshape((X.shape[0], -1))
-        sklearn_X = cdist(X_, X_, metric="euclidean")
-    else:
-        X_ = to_time_series_dataset(X)
-        n, sz, d = X_.shape
-        sklearn_X = X_.reshape((n, -1))
-
-        def sklearn_metric(x, y):
-            return metric(to_time_series(x.reshape((sz, d)),
-                                         remove_nans=True),
-                          to_time_series(y.reshape((sz, d)),
-                                         remove_nans=True))
-    metric = "precomputed" if sklearn_metric is None else sklearn_metric
-    return sklearn_silhouette_score(X=sklearn_X,
-                                    labels=labels,
-                                    metric=metric,
-                                    sample_size=sample_size,
-                                    random_state=random_state,
-                                    **kwds)
-
-
-def _check_initial_guess(init, n_clusters):
-    if hasattr(init, '__array__'):
-        assert init.shape[0] == n_clusters, \
-            "Initial guess index array must contain {} samples," \
-            " {} given".format(n_clusters, init.shape[0])
 
 
 class KernelKMeans(ClusterMixin, BaseModelPackage, TimeSeriesBaseEstimator):
@@ -654,18 +450,6 @@ class GlobalAlignmentKernelKMeans(KernelKMeans):
             DeprecationWarning, stacklevel=2)
         super().__init__(**kwargs)
         self.kernel = "gak"
-
-
-class TimeSeriesCentroidBasedClusteringMixin:
-    """Mixin class for centroid-based clustering of time series."""
-    def _post_fit(self, X_fitted, centroids, inertia):
-        if numpy.isfinite(inertia) and (centroids is not None):
-            self.cluster_centers_ = centroids
-            self._assign(X_fitted)
-            self._X_fit = X_fitted
-            self.inertia_ = inertia
-        else:
-            self._X_fit = None
 
 
 class TimeSeriesKMeans(TransformerMixin, ClusterMixin,
@@ -1057,276 +841,3 @@ class TimeSeriesKMeans(TransformerMixin, ClusterMixin,
 
     def _more_tags(self):
         return {'allow_nan': True, 'allow_variable_length': True}
-
-
-class KShape(ClusterMixin, TimeSeriesCentroidBasedClusteringMixin,
-             BaseModelPackage, TimeSeriesBaseEstimator):
-    """KShape clustering for time series.
-
-    KShape was originally presented in [1]_.
-
-    Parameters
-    ----------
-    n_clusters : int (default: 3)
-        Number of clusters to form.
-
-    max_iter : int (default: 100)
-        Maximum number of iterations of the k-Shape algorithm.
-
-    tol : float (default: 1e-6)
-        Inertia variation threshold. If at some point, inertia varies less than
-        this threshold between two consecutive
-        iterations, the model is considered to have converged and the algorithm
-        stops.
-
-    n_init : int (default: 1)
-        Number of time the k-Shape algorithm will be run with different
-        centroid seeds. The final results will be the
-        best output of n_init consecutive runs in terms of inertia.
-
-    verbose : bool (default: False)
-        Whether or not to print information about the inertia while learning
-        the model.
-
-    random_state : integer or numpy.RandomState, optional
-        Generator used to initialize the centers. If an integer is given, it
-        fixes the seed. Defaults to the global
-        numpy random number generator.
-
-    init : {'random' or ndarray} (default: 'random')
-        Method for initialization.
-        'random': choose k observations (rows) at random from data for the
-        initial centroids.
-        If an ndarray is passed, it should be of shape (n_clusters, ts_size, d)
-        and gives the initial centers.
-
-    Attributes
-    ----------
-    cluster_centers_ : numpy.ndarray of shape (sz, d).
-        Centroids
-
-    labels_ : numpy.ndarray of integers with shape (n_ts, ).
-        Labels of each point
-
-    inertia_ : float
-        Sum of distances of samples to their closest cluster center.
-
-    n_iter_ : int
-        The number of iterations performed during fit.
-
-    Notes
-    -----
-        This method requires a dataset of equal-sized time series.
-
-    Examples
-    --------
-    >>> from tslearn.generators import random_walks
-    >>> X = random_walks(n_ts=50, sz=32, d=1)
-    >>> X = TimeSeriesScalerMeanVariance(mu=0., std=1.).fit_transform(X)
-    >>> ks = KShape(n_clusters=3, n_init=1, random_state=0).fit(X)
-    >>> ks.cluster_centers_.shape
-    (3, 32, 1)
-
-    References
-    ----------
-    .. [1] J. Paparrizos & L. Gravano. k-Shape: Efficient and Accurate
-       Clustering of Time Series. SIGMOD 2015. pp. 1855-1870.
-    """
-
-    def __init__(self, n_clusters=3, max_iter=100, tol=1e-6, n_init=1,
-                 verbose=False, random_state=None, init='random'):
-        self.n_clusters = n_clusters
-        self.max_iter = max_iter
-        self.tol = tol
-        self.random_state = random_state
-        self.n_init = n_init
-        self.verbose = verbose
-        self.init = init
-
-    def _is_fitted(self):
-        """
-        Check if the model has been fit.
-
-        Returns
-        -------
-        bool
-        """
-
-        check_is_fitted(self,
-                        ['cluster_centers_', 'norms_', 'norms_centroids_'])
-        return True
-
-    def _shape_extraction(self, X, k):
-        sz = X.shape[1]
-        Xp = y_shifted_sbd_vec(self.cluster_centers_[k], X[self.labels_ == k],
-                               norm_ref=-1,
-                               norms_dataset=self.norms_[self.labels_ == k])
-        S = numpy.dot(Xp[:, :, 0].T, Xp[:, :, 0])
-        Q = numpy.eye(sz) - numpy.ones((sz, sz)) / sz
-        M = numpy.dot(Q.T, numpy.dot(S, Q))
-        _, vec = numpy.linalg.eigh(M)
-        mu_k = vec[:, -1].reshape((sz, 1))
-
-        # The way the optimization problem is (ill-)formulated, both mu_k and
-        # -mu_k are candidates for barycenters
-        # In the following, we check which one is best candidate
-        dist_plus_mu = numpy.sum(numpy.linalg.norm(Xp - mu_k, axis=(1, 2)))
-        dist_minus_mu = numpy.sum(numpy.linalg.norm(Xp + mu_k, axis=(1, 2)))
-        if dist_minus_mu < dist_plus_mu:
-            mu_k *= -1
-
-        return mu_k
-
-    def _update_centroids(self, X):
-        for k in range(self.n_clusters):
-            self.cluster_centers_[k] = self._shape_extraction(X, k)
-        self.cluster_centers_ = TimeSeriesScalerMeanVariance(
-            mu=0., std=1.).fit_transform(self.cluster_centers_)
-        self.norms_centroids_ = numpy.linalg.norm(self.cluster_centers_,
-                                                  axis=(1, 2))
-
-    def _cross_dists(self, X):
-        return 1. - cdist_normalized_cc(X, self.cluster_centers_,
-                                        norms1=self.norms_,
-                                        norms2=self.norms_centroids_,
-                                        self_similarity=False)
-
-    def _assign(self, X):
-        dists = self._cross_dists(X)
-        self.labels_ = dists.argmin(axis=1)
-        _check_no_empty_cluster(self.labels_, self.n_clusters)
-        self.inertia_ = _compute_inertia(dists, self.labels_)
-
-    def _fit_one_init(self, X, rs):
-        if hasattr(self.init, '__array__'):
-            self.cluster_centers_ = self.init.copy()
-        elif self.init == "random":
-            indices = rs.choice(X.shape[0], self.n_clusters)
-            self.cluster_centers_ = X[indices].copy()
-        else:
-            raise ValueError("Value %r for parameter 'init' is "
-                             "invalid" % self.init)
-        self.norms_centroids_ = numpy.linalg.norm(self.cluster_centers_,
-                                                  axis=(1, 2))
-        self._assign(X)
-        old_inertia = numpy.inf
-
-        for it in range(self.max_iter):
-            old_cluster_centers = self.cluster_centers_.copy()
-            self._update_centroids(X)
-            self._assign(X)
-            if self.verbose:
-                print("%.3f" % self.inertia_, end=" --> ")
-
-            if numpy.abs(old_inertia - self.inertia_) < self.tol or \
-                    (old_inertia - self.inertia_ < 0):
-                self.cluster_centers_ = old_cluster_centers
-                self._assign(X)
-                break
-
-            old_inertia = self.inertia_
-        if self.verbose:
-            print("")
-
-        self._iter = it + 1
-
-        return self
-
-    def fit(self, X, y=None):
-        """Compute k-Shape clustering.
-
-        Parameters
-        ----------
-        X : array-like of shape=(n_ts, sz, d)
-            Time series dataset.
-
-        y
-            Ignored
-        """
-        X = check_array(X, allow_nd=True)
-
-        max_attempts = max(self.n_init, 10)
-
-        self.labels_ = None
-        self.inertia_ = numpy.inf
-        self.cluster_centers_ = None
-
-        self.norms_ = 0.
-        self.norms_centroids_ = 0.
-
-        self.n_iter_ = 0
-
-        X_ = to_time_series_dataset(X)
-        self._X_fit = X_
-        self.norms_ = numpy.linalg.norm(X_, axis=(1, 2))
-
-        _check_initial_guess(self.init, self.n_clusters)
-
-        rs = check_random_state(self.random_state)
-
-        best_correct_centroids = None
-        min_inertia = numpy.inf
-        n_successful = 0
-        n_attempts = 0
-        while n_successful < self.n_init and n_attempts < max_attempts:
-            try:
-                if self.verbose and self.n_init > 1:
-                    print("Init %d" % (n_successful + 1))
-                n_attempts += 1
-                self._fit_one_init(X_, rs)
-                if self.inertia_ < min_inertia:
-                    best_correct_centroids = self.cluster_centers_.copy()
-                    min_inertia = self.inertia_
-                    self.n_iter_ = self._iter
-                n_successful += 1
-            except EmptyClusterError:
-                if self.verbose:
-                    print("Resumed because of empty cluster")
-        self.norms_centroids_ = numpy.linalg.norm(self.cluster_centers_,
-                                                  axis=(1, 2))
-        self._post_fit(X_, best_correct_centroids, min_inertia)
-        return self
-
-    def fit_predict(self, X, y=None):
-        """Fit k-Shape clustering using X and then predict the closest cluster
-        each time series in X belongs to.
-
-        It is more efficient to use this method than to sequentially call fit
-        and predict.
-
-        Parameters
-        ----------
-        X : array-like of shape=(n_ts, sz, d)
-            Time series dataset to predict.
-
-        y
-            Ignored
-
-        Returns
-        -------
-        labels : array of shape=(n_ts, )
-            Index of the cluster each sample belongs to.
-        """
-        return self.fit(X, y).labels_
-
-    def predict(self, X):
-        """Predict the closest cluster each time series in X belongs to.
-
-        Parameters
-        ----------
-        X : array-like of shape=(n_ts, sz, d)
-            Time series dataset to predict.
-
-        Returns
-        -------
-        labels : array of shape=(n_ts, )
-            Index of the cluster each sample belongs to.
-        """
-        X = check_array(X, allow_nd=True)
-        check_is_fitted(self,
-                        ['cluster_centers_', 'norms_', 'norms_centroids_'])
-
-        X_ = check_dims(X, X_fit_dims=self.cluster_centers_.shape)
-        X_ = TimeSeriesScalerMeanVariance(mu=0., std=1.).fit_transform(X_)
-        dists = self._cross_dists(X_)
-        return dists.argmin(axis=1)
